Completely rewrite the compiler using Java
Yes, I completely rewrote the compiler! While the Python implementation
had a really well written (IMO) parser, the compiler was a mess, and all
in all it was difficult to extend. With this new implementation, adding
new functionality to the language and the compiler will be much less of
a pain!